Move back the ContextQuery related extensions and corresponding SyntaxFact APIs up to the shared WorkspaceExtensions project
