+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Extensions/SyntaxTreeExtensions.cs
@@ -4,14 +4,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
-using Microsoft.CodeAnalysis.Text;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Extensions
 {
@@ -156,66 +152,12 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             return token.Kind() == SyntaxKind.NumericLiteralToken;
         }
 
-        public static bool IsPrimaryFunctionExpressionContext(this SyntaxTree syntaxTree, int position, SyntaxToken tokenOnLeftOfPosition)
-        {
-            return
-                syntaxTree.IsTypeOfExpressionContext(position, tokenOnLeftOfPosition) ||
-                syntaxTree.IsDefaultExpressionContext(position, tokenOnLeftOfPosition) ||
-                syntaxTree.IsSizeOfExpressionContext(position, tokenOnLeftOfPosition);
-        }
-
-        public static bool IsTypeOfExpressionContext(this SyntaxTree syntaxTree, int position, SyntaxToken tokenOnLeftOfPosition)
-        {
-            var token = tokenOnLeftOfPosition.GetPreviousTokenIfTouchingWord(position);
-
-            if (token.IsKind(SyntaxKind.OpenParenToken) && token.Parent.IsKind(SyntaxKind.TypeOfExpression))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        public static bool IsDefaultExpressionContext(this SyntaxTree syntaxTree, int position, SyntaxToken tokenOnLeftOfPosition)
-        {
-            var token = tokenOnLeftOfPosition.GetPreviousTokenIfTouchingWord(position);
-
-            if (token.IsKind(SyntaxKind.OpenParenToken) && token.Parent.IsKind(SyntaxKind.DefaultExpression))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        public static bool IsSizeOfExpressionContext(
-            this SyntaxTree syntaxTree, int position, SyntaxToken tokenOnLeftOfPosition)
-        {
-            var token = tokenOnLeftOfPosition.GetPreviousTokenIfTouchingWord(position);
-
-            if (token.IsKind(SyntaxKind.OpenParenToken) && token.Parent.IsKind(SyntaxKind.SizeOfExpression))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
         public static bool IsAfterKeyword(this SyntaxTree syntaxTree, int position, SyntaxKind kind, CancellationToken cancellationToken)
         {
             var token = syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken);
             token = token.GetPreviousTokenIfTouchingWord(position);
 
             return token.Kind() == kind;
-        }
-
-        public static bool IsInNonUserCode(this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            return
-                syntaxTree.IsEntirelyWithinNonUserCodeComment(position, cancellationToken) ||
-                syntaxTree.IsEntirelyWithinConflictMarker(position, cancellationToken) ||
-                syntaxTree.IsEntirelyWithinStringOrCharLiteral(position, cancellationToken) ||
-                syntaxTree.IsInInactiveRegion(position, cancellationToken);
         }
 
         public static bool IsEntirelyWithinNonUserCodeComment(this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
@@ -488,437 +430,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             }
 
             return false;
-        }
-
-        public static bool IsInInactiveRegion(
-            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            Contract.ThrowIfNull(syntaxTree);
-
-            // cases:
-            // $ is EOF
-
-            // #if false
-            //    |
-
-            // #if false
-            //    |$
-
-            // #if false
-            // |
-
-            // #if false
-            // |$
-
-            if (syntaxTree.IsPreProcessorKeywordContext(position, cancellationToken))
-            {
-                return false;
-            }
-
-            // The latter two are the hard cases we don't actually have an 
-            // DisabledTextTrivia yet. 
-            var trivia = syntaxTree.GetRoot(cancellationToken).FindTrivia(position, findInsideTrivia: false);
-            if (trivia.Kind() == SyntaxKind.DisabledTextTrivia)
-            {
-                return true;
-            }
-
-            var token = syntaxTree.FindTokenOrEndToken(position, cancellationToken);
-            if (token.Kind() == SyntaxKind.EndOfFileToken)
-            {
-                var triviaList = token.LeadingTrivia;
-                foreach (var triviaTok in triviaList.Reverse())
-                {
-                    if (triviaTok.Span.Contains(position))
-                    {
-                        return false;
-                    }
-
-                    if (triviaTok.Span.End < position)
-                    {
-                        if (!triviaTok.HasStructure)
-                        {
-                            return false;
-                        }
-
-                        var structure = triviaTok.GetStructure();
-                        if (structure is BranchingDirectiveTriviaSyntax branch)
-                        {
-                            return !branch.IsActive || !branch.BranchTaken;
-                        }
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        public static ImmutableArray<MemberDeclarationSyntax> GetFieldsAndPropertiesInSpan(
-            this SyntaxNode root, TextSpan textSpan, bool allowPartialSelection)
-        {
-            var token = root.FindTokenOnRightOfPosition(textSpan.Start);
-            var firstMember = token.GetAncestors<MemberDeclarationSyntax>().FirstOrDefault();
-            if (firstMember != null)
-            {
-                if (firstMember.Parent is TypeDeclarationSyntax containingType)
-                {
-                    return GetFieldsAndPropertiesInSpan(textSpan, containingType, firstMember, allowPartialSelection);
-                }
-            }
-
-            return ImmutableArray<MemberDeclarationSyntax>.Empty;
-        }
-
-        private static ImmutableArray<MemberDeclarationSyntax> GetFieldsAndPropertiesInSpan(
-            TextSpan textSpan,
-            TypeDeclarationSyntax containingType,
-            MemberDeclarationSyntax firstMember,
-            bool allowPartialSelection)
-        {
-            var members = containingType.Members;
-            var fieldIndex = members.IndexOf(firstMember);
-            if (fieldIndex < 0)
-            {
-                return ImmutableArray<MemberDeclarationSyntax>.Empty;
-            }
-
-            var selectedMembers = ArrayBuilder<MemberDeclarationSyntax>.GetInstance();
-            for (var i = fieldIndex; i < members.Count; i++)
-            {
-                var member = members[i];
-                if (IsSelectedFieldOrProperty(textSpan, member, allowPartialSelection))
-                {
-                    selectedMembers.Add(member);
-                }
-            }
-
-            return selectedMembers.ToImmutableAndFree();
-
-            // local functions
-            static bool IsSelectedFieldOrProperty(TextSpan textSpan, MemberDeclarationSyntax member, bool allowPartialSelection)
-            {
-                if (!member.IsKind(SyntaxKind.FieldDeclaration, SyntaxKind.PropertyDeclaration))
-                {
-                    return false;
-                }
-
-                // first, check if entire member is selected
-                if (textSpan.Contains(member.Span))
-                {
-                    return true;
-                }
-
-                if (!allowPartialSelection)
-                {
-                    return false;
-                }
-
-                // next, check if identifier is at least partially selected
-                switch (member)
-                {
-                    case FieldDeclarationSyntax field:
-                        var variables = field.Declaration.Variables;
-                        foreach (var variable in variables)
-                        {
-                            if (textSpan.OverlapsWith(variable.Identifier.Span))
-                            {
-                                return true;
-                            }
-                        }
-                        return false;
-                    case PropertyDeclarationSyntax property:
-                        return textSpan.OverlapsWith(property.Identifier.Span);
-                    default:
-                        return false;
-                }
-            }
-        }
-
-        public static bool IsInPartiallyWrittenGeneric(
-            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            return syntaxTree.IsInPartiallyWrittenGeneric(position, cancellationToken, out var genericIdentifier, out var lessThanToken);
-        }
-
-        public static bool IsInPartiallyWrittenGeneric(
-            this SyntaxTree syntaxTree,
-            int position,
-            CancellationToken cancellationToken,
-            out SyntaxToken genericIdentifier)
-        {
-            return syntaxTree.IsInPartiallyWrittenGeneric(position, cancellationToken, out genericIdentifier, out var lessThanToken);
-        }
-
-        public static bool IsInPartiallyWrittenGeneric(
-            this SyntaxTree syntaxTree,
-            int position,
-            CancellationToken cancellationToken,
-            out SyntaxToken genericIdentifier,
-            out SyntaxToken lessThanToken)
-        {
-            genericIdentifier = default;
-            lessThanToken = default;
-            var index = 0;
-
-            var token = syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken);
-            if (token.Kind() == SyntaxKind.None)
-            {
-                return false;
-            }
-
-            // check whether we are under type or member decl
-            if (token.GetAncestor<TypeParameterListSyntax>() != null)
-            {
-                return false;
-            }
-
-            var stack = 0;
-            while (true)
-            {
-                switch (token.Kind())
-                {
-                    case SyntaxKind.LessThanToken:
-                        if (stack == 0)
-                        {
-                            // got here so we read successfully up to a < now we have to read the
-                            // name before that and we're done!
-                            lessThanToken = token;
-                            token = token.GetPreviousToken(includeSkipped: true);
-                            if (token.Kind() == SyntaxKind.None)
-                            {
-                                return false;
-                            }
-
-                            // ok
-                            // so we've read something like:
-                            // ~~~~~~~~~<a,b,...
-                            // but we need to know the simple name that precedes the <
-                            // it could be
-                            // ~~~~~~goo<a,b,...
-                            if (token.Kind() == SyntaxKind.IdentifierToken)
-                            {
-                                // okay now check whether it is actually partially written
-                                if (IsFullyWrittenGeneric(token, lessThanToken))
-                                {
-                                    return false;
-                                }
-
-                                genericIdentifier = token;
-                                return true;
-                            }
-
-                            return false;
-                        }
-                        else
-                        {
-                            stack--;
-                            break;
-                        }
-
-                    case SyntaxKind.GreaterThanGreaterThanToken:
-                        stack++;
-                        goto case SyntaxKind.GreaterThanToken;
-
-                    // fall through
-                    case SyntaxKind.GreaterThanToken:
-                        stack++;
-                        break;
-
-                    case SyntaxKind.AsteriskToken:      // for int*
-                    case SyntaxKind.QuestionToken:      // for int?
-                    case SyntaxKind.ColonToken:         // for global::  (so we don't dismiss help as you type the first :)
-                    case SyntaxKind.ColonColonToken:    // for global::
-                    case SyntaxKind.CloseBracketToken:
-                    case SyntaxKind.OpenBracketToken:
-                    case SyntaxKind.DotToken:
-                    case SyntaxKind.IdentifierToken:
-                        break;
-
-                    case SyntaxKind.CommaToken:
-                        if (stack == 0)
-                        {
-                            index++;
-                        }
-
-                        break;
-
-                    default:
-                        // user might have typed "in" on the way to typing "int"
-                        // don't want to disregard this genericname because of that
-                        if (SyntaxFacts.IsKeywordKind(token.Kind()))
-                        {
-                            break;
-                        }
-
-                        // anything else and we're sunk.
-                        return false;
-                }
-
-                // look backward one token, include skipped tokens, because the parser frequently
-                // does skip them in cases like: "Func<A, B", which get parsed as: expression
-                // statement "Func<A" with missing semicolon, expression statement "B" with missing
-                // semicolon, and the "," is skipped.
-                token = token.GetPreviousToken(includeSkipped: true);
-                if (token.Kind() == SyntaxKind.None)
-                {
-                    return false;
-                }
-            }
-        }
-
-        private static bool IsFullyWrittenGeneric(SyntaxToken token, SyntaxToken lessThanToken)
-        {
-            return token.Parent is GenericNameSyntax genericName && genericName.TypeArgumentList != null &&
-                   genericName.TypeArgumentList.LessThanToken == lessThanToken && !genericName.TypeArgumentList.GreaterThanToken.IsMissing;
-        }
-
-        public static bool IsPreProcessorDirectiveContext(this SyntaxTree syntaxTree, int position, SyntaxToken preProcessorTokenOnLeftOfPosition, CancellationToken cancellationToken)
-        {
-            var token = preProcessorTokenOnLeftOfPosition;
-            var directive = token.GetAncestor<DirectiveTriviaSyntax>();
-
-            // Directives contain the EOL, so if the position is within the full span of the
-            // directive, then it is on that line, the only exception is if the directive is on the
-            // last line, the position at the end if technically not contained by the directive but
-            // its also not on a new line, so it should be considered part of the preprocessor
-            // context.
-            if (directive == null)
-            {
-                return false;
-            }
-
-            return
-                directive.FullSpan.Contains(position) ||
-                directive.FullSpan.End == syntaxTree.GetRoot(cancellationToken).FullSpan.End;
-        }
-
-        public static bool IsPreProcessorDirectiveContext(this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            var leftToken = syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken, includeDirectives: true);
-
-            return syntaxTree.IsPreProcessorDirectiveContext(position, leftToken, cancellationToken);
-        }
-
-        public static bool IsPreProcessorKeywordContext(this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            return IsPreProcessorKeywordContext(
-                syntaxTree, position,
-                syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken, includeDirectives: true));
-        }
-
-        public static bool IsPreProcessorKeywordContext(this SyntaxTree syntaxTree, int position, SyntaxToken preProcessorTokenOnLeftOfPosition)
-        {
-            // cases:
-            //  #|
-            //  #d|
-            //  # |
-            //  # d|
-
-            // note: comments are not allowed between the # and item.
-            var token = preProcessorTokenOnLeftOfPosition;
-            token = token.GetPreviousTokenIfTouchingWord(position);
-
-            if (token.IsKind(SyntaxKind.HashToken))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Are you possibly typing a tuple type or expression?
-        /// This is used to suppress colon as a completion trigger (so that you can type element names).
-        /// This is also used to recommend some keywords (like var).
-        /// </summary>
-        public static bool IsPossibleTupleContext(this SyntaxTree syntaxTree, SyntaxToken leftToken, int position)
-        {
-            leftToken = leftToken.GetPreviousTokenIfTouchingWord(position);
-
-            // ($$
-            // (a, $$
-            if (IsPossibleTupleOpenParenOrComma(leftToken))
-            {
-                return true;
-            }
-
-            // ((a, b) $$
-            // (..., (a, b) $$
-            if (leftToken.IsKind(SyntaxKind.CloseParenToken))
-            {
-                if (leftToken.Parent.IsKind(
-                        SyntaxKind.ParenthesizedExpression,
-                        SyntaxKind.TupleExpression,
-                        SyntaxKind.TupleType))
-                {
-                    var possibleCommaOrParen = FindTokenOnLeftOfNode(leftToken.Parent);
-                    if (IsPossibleTupleOpenParenOrComma(possibleCommaOrParen))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            // (a $$
-            // (..., b $$
-            if (leftToken.IsKind(SyntaxKind.IdentifierToken))
-            {
-                var possibleCommaOrParen = FindTokenOnLeftOfNode(leftToken.Parent);
-                if (IsPossibleTupleOpenParenOrComma(possibleCommaOrParen))
-                {
-                    return true;
-                }
-            }
-
-            // (a.b $$
-            // (..., a.b $$
-            if (leftToken.IsKind(SyntaxKind.IdentifierToken) &&
-                leftToken.Parent.IsKind(SyntaxKind.IdentifierName) &&
-                leftToken.Parent.IsParentKind(SyntaxKind.QualifiedName, SyntaxKind.SimpleMemberAccessExpression))
-            {
-                var possibleCommaOrParen = FindTokenOnLeftOfNode(leftToken.Parent.Parent);
-                if (IsPossibleTupleOpenParenOrComma(possibleCommaOrParen))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        public static bool IsPossibleTupleOpenParenOrComma(this SyntaxToken possibleCommaOrParen)
-        {
-            if (!possibleCommaOrParen.IsKind(SyntaxKind.OpenParenToken, SyntaxKind.CommaToken))
-            {
-                return false;
-            }
-
-            if (possibleCommaOrParen.Parent.IsKind(
-                    SyntaxKind.ParenthesizedExpression,
-                    SyntaxKind.TupleExpression,
-                    SyntaxKind.TupleType,
-                    SyntaxKind.CastExpression))
-            {
-                return true;
-            }
-
-            // in script
-            if (possibleCommaOrParen.Parent.IsKind(SyntaxKind.ParameterList) &&
-                possibleCommaOrParen.Parent.IsParentKind(SyntaxKind.ParenthesizedLambdaExpression))
-            {
-                var parenthesizedLambda = (ParenthesizedLambdaExpressionSyntax)possibleCommaOrParen.Parent.Parent;
-                if (parenthesizedLambda.ArrowToken.IsMissing)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private static SyntaxToken FindTokenOnLeftOfNode(SyntaxNode node)
-        {
-            return node.FindTokenOnLeftOfPosition(node.SpanStart);
         }
     }
 }
