+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/VisualBasic/Extensions/SyntaxTreeExtensions.vb
@@ -437,23 +437,5 @@ recurse:
         Public Function GetFirstEnclosingStatement(node As SyntaxNode) As StatementSyntax
             Return node.AncestorsAndSelf().Where(Function(n) TypeOf n Is StatementSyntax).OfType(Of StatementSyntax).FirstOrDefault()
         End Function
-
-        ' Tuple literals aren't recognized by the parser until there is a comma
-        ' So a parenthesized expression is a possible tuple context too
-        <Extension>
-        Friend Function IsPossibleTupleContext(syntaxTree As SyntaxTree,
-                                               tokenOnLeftOfPosition As SyntaxToken,
-                                               position As Integer) As Boolean
-
-            tokenOnLeftOfPosition = tokenOnLeftOfPosition.GetPreviousTokenIfTouchingWord(position)
-
-            If tokenOnLeftOfPosition.IsKind(SyntaxKind.OpenParenToken) Then
-                Return tokenOnLeftOfPosition.Parent.IsKind(SyntaxKind.ParenthesizedExpression,
-                                                           SyntaxKind.TupleExpression, SyntaxKind.TupleType)
-            End If
-
-            Return tokenOnLeftOfPosition.IsKind(SyntaxKind.CommaToken) AndAlso
-                tokenOnLeftOfPosition.Parent.IsKind(SyntaxKind.TupleExpression, SyntaxKind.TupleType)
-        End Function
     End Module
 End Namespace
